hornet/backend/requirements.txt
django==4.0.2
django-ninja==0.17.0